+++ client/src/pages/board/[id].tsx
@@ -23,6 +23,8 @@ export default function PostDetailPage({ pageProps }: IPostDetailPage) {
   const data = pageProps.data;
   const session = useSession();
   const user = session.data?.user;
+  const comments = data.comments as IComment[] | null;
+
   return (
     <Layout>
       <div className="flex flex-col max-w-4xl m-auto gap-14 mb-4">
@@ -30,7 +32,7 @@ export default function PostDetailPage({ pageProps }: IPostDetailPage) {
         <h3 className="text-2xl font-bold pl-3">댓글 3</h3>
         <div className="flex flex-col border-gray-100 border-[1px] border-collapse">
           <CommentUpload user={user} />
-          {data.comments?.map((comment: any) => (
+          {comments?.map((comment: IComment) => (
             <Comment key={comment.commentOriginId} data={comment} />
           ))}
         </div>
